requirements.txt
# Lateryx Security Analyzer - Python Dependencies
# =================================================

# Core graph analysis library
networkx>=3.2.1

# Terraform/HCL parsing and security scanning
checkov>=3.1.0

# Configuration loading
pyyaml>=6.0.1

# Cloud providers (optional - uncomment as needed)
boto3>=1.28.0        # AWS integration
# azure-identity>=1.14  # Azure
# google-cloud-core     # GCP

# CLI and utilities
click>=8.1.7

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0

# Type checking
mypy>=1.7.1
